Reverts folder rename
