+++ apps/judicial-system/web/pages/verjandi/[id].ts
@@ -1,3 +1,3 @@
-import { SignedVerdictOverview } from '@island.is/judicial-system-web/src/routes/Shared/CaseOverview/SignedVerdictOverview'
+import { SignedVerdictOverview } from '@island.is/judicial-system-web/src/routes/Shared/SignedVerdictOverview/SignedVerdictOverview'
 
 export default SignedVerdictOverview
