+++ apps/judicial-system/web/src/routes/Shared/SignedVerdictOverview/Components/AppealSection/AppealSection.css.ts
