+++ apps/judicial-system/web/src/routes/Shared/SignedVerdictOverview/Components/SignedDocument.tsx
