+++ apps/judicial-system/web/src/routes/Shared/SignedVerdictOverview/courtRecordSignatureConfirmationGql.ts
